requirements.txt
aiomysql==0.2.0
discord.py==2.3.1
sqlalchemy==1.3.0